Players should now be able to beat the first three dungeons in any order they so choose
The Kokiri blocking the exit of the Kokiri Forest now stands guard next to it, allowing Link to leave early. Saria, Fado, Mido, and the rest of the Kokiri have had their behaviour updated to compensate for this possibility.

Mido allows you to pass if you have the Hylian Shield equipped

The boulders blocking the Zora River are gone to allow players to enter Jabu Jabu without the bomb bag.

Alternate methods of solving the Dodongo fire ring rooms have been added, along with a new property of the fire rings to allow players to beat the dungeon without the Slingshot.

The Zora Waterfall is always open, such that Hyrule Castle is a prerequisite only for the Goron's Ruby. The waterfall still has the text saying that the waterfall slumbers when the king does, but I interpret this as being a boast that the king never rests. Also Kaepora Gaebora no longer appears to tell you to play Zelda's Lullaby in front of the Zora Waterfall.

A new method of solving Jabu Jabu's mouth has been added so players do not need the slingshot to beat the dungeon