Players should now be able to beat the final five dungeons in any order they so choose
Added a bombchu bridge in shadow temple so players can reach the boss door without the bow or being an ace bombchu sniper

Removed enemies and a rather pointless eye switch puzzle in spirit temple so players can get the boss key without the bow or using damage invulnerability exploits

Added two boxes in the Haunted Wastelands so players can reach the Spirit Temple without the Hookshot

Bongo Bongo now escapes from the Shadow Temple the moment you enter Kakariko Village, allowing players to reach the Shadow Temple without completing the Forest, Fire, and Water Temples

Removed the lock on the first floor entrance to the central pillar room in the Water Temple and removed a rather pointless eye switch puzzle in water temple so players can get the boss key without the bow or using warps to skip using keys. As a result, there's 1 key leftover that's only accessible with the bow, consider it a method of reducing the dungeon's difficulty by doing the forest temple first.

The rainbow bridge to ganondorf's castle requires the water, fire, and forest medallions in addition to the spirit and shadow medallions